+++ openeo/rest/connection.py
@@ -21,6 +21,7 @@ from typing import (
     List,
     Optional,
     Sequence,
+    Set,
     Tuple,
     Union,
 )
@@ -53,7 +54,7 @@ from openeo.rest import (
     OpenEoClientException,
     OpenEoRestError,
 )
-from openeo.rest._datacube import build_child_callback
+from openeo.rest._datacube import _ProcessGraphAbstraction, build_child_callback
 from openeo.rest.auth.auth import BasicBearerAuth, BearerAuth, NullAuth, OidcBearerAuth
 from openeo.rest.auth.config import AuthConfig, RefreshTokenStore
 from openeo.rest.auth.oidc import (
@@ -1128,11 +1129,19 @@ class Connection(RestApiConnection):
         """
         return RESTUserDefinedProcess(user_defined_process_id=user_defined_process_id, connection=self)
 
-    def validate_process_graph(self, process_graph: Union[dict, FlatGraphableMixin, Any]) -> List[dict]:
+    def validate_process_graph(
+        self, process_graph: Union[dict, FlatGraphableMixin, str, Path, List[FlatGraphableMixin]]
+    ) -> List[dict]:
         """
         Validate a process graph without executing it.
 
-        :param process_graph: (flat) dict representing process graph
+        :param process_graph: openEO-style (flat) process graph representation,
+            or an object that can be converted to such a representation:
+            a dictionary, a :py:class:`~openeo.rest.datacube.DataCube` object,
+            a string with a JSON representation,
+            a local file path or URL to a JSON representation,
+            a :py:class:`~openeo.rest.multiresult.MultiResult` object, ...
+
         :return: list of errors (dictionaries with "code" and "message" fields)
         """
         pg_with_metadata = self._build_request_with_process_graph(process_graph)["process"]
@@ -1608,12 +1617,19 @@ class Connection(RestApiConnection):
             metadata = resp.json()
         return UserFile.from_metadata(metadata=metadata, connection=self)
 
-    def _build_request_with_process_graph(self, process_graph: Union[dict, FlatGraphableMixin, Any], **kwargs) -> dict:
+    def _build_request_with_process_graph(
+        self,
+        process_graph: Union[dict, FlatGraphableMixin, str, Path, List[FlatGraphableMixin]],
+        **kwargs,
+    ) -> dict:
         """
         Prepare a json payload with a process graph to submit to /result, /services, /jobs, ...
         :param process_graph: flat dict representing a "process graph with metadata" ({"process": {"process_graph": ...}, ...})
         """
         # TODO: make this a more general helper (like `as_flat_graph`)
+        connections = extract_connections(process_graph)
+        if any(c != self for c in connections):
+            raise OpenEoClientException(f"Mixing different connections: {self} and {connections}.")
         result = kwargs
         process_graph = as_flat_graph(process_graph)
         if "process_graph" not in process_graph:
@@ -1656,7 +1672,7 @@ class Connection(RestApiConnection):
     # TODO: unify `download` and `execute` better: e.g. `download` always writes to disk, `execute` returns result (raw or as JSON decoded dict)
     def download(
         self,
-        graph: Union[dict, FlatGraphableMixin, str, Path],
+        graph: Union[dict, FlatGraphableMixin, str, Path, List[FlatGraphableMixin]],
         outputfile: Union[Path, str, None] = None,
         *,
         timeout: Optional[int] = None,
@@ -1695,7 +1711,7 @@ class Connection(RestApiConnection):
 
     def execute(
         self,
-        process_graph: Union[dict, str, Path],
+        process_graph: Union[dict, FlatGraphableMixin, str, Path, List[FlatGraphableMixin]],
         *,
         timeout: Optional[int] = None,
         validate: Optional[bool] = None,
@@ -1732,7 +1748,7 @@ class Connection(RestApiConnection):
 
     def create_job(
         self,
-        process_graph: Union[dict, str, Path, FlatGraphableMixin],
+        process_graph: Union[dict, FlatGraphableMixin, str, Path, List[FlatGraphableMixin]],
         *,
         title: Optional[str] = None,
         description: Optional[str] = None,
@@ -1744,8 +1760,12 @@ class Connection(RestApiConnection):
         """
         Create a new job from given process graph on the back-end.
 
-        :param process_graph: (flat) dict representing a process graph, or process graph as raw JSON string,
-            or as local file path or URL
+        :param process_graph: openEO-style (flat) process graph representation,
+            or an object that can be converted to such a representation:
+            a dictionary, a :py:class:`~openeo.rest.datacube.DataCube` object,
+            a string with a JSON representation,
+            a local file path or URL to a JSON representation,
+            a :py:class:`~openeo.rest.multiresult.MultiResult` object, ...
         :param title: job title
         :param description: job description
         :param plan: The billing plan to process and charge the job with
@@ -1755,6 +1775,9 @@ class Connection(RestApiConnection):
         :param validate: Optional toggle to enable/prevent validation of the process graphs before execution
             (overruling the connection's ``auto_validate`` setting).
         :return: Created job
+
+        .. versionchanged:: 0.35.0
+            Add :ref:`multi-result support <multi-result-process-graphs>`.
         """
         # TODO move all this (BatchJob factory) logic to BatchJob?
 
@@ -1968,3 +1991,25 @@ def paginate(con: Connection, url: str, params: Optional[dict] = None, callback:
         url = next_links[0]["href"]
         page += 1
         params = {}
+
+
+def extract_connections(
+    data: Union[_ProcessGraphAbstraction, Sequence[_ProcessGraphAbstraction], Any]
+) -> Set[Connection]:
+    """
+    Extract the :py:class:`Connection` object(s) linked from a given data construct.
+    Typical use case is to get the connection from a :py:class:`DataCube`,
+    but can also extract multiple connections from a list of data cubes.
+    """
+    connections = set()
+    # TODO: define some kind of "Connected" interface/mixin/protocol
+    #       for objects that contain a connection instead of just checking for _ProcessGraphAbstraction
+    # TODO: also support extracting connections from other objects like BatchJob, ...
+    if isinstance(data, _ProcessGraphAbstraction) and data.connection:
+        connections.add(data.connection)
+    elif isinstance(data, (list, tuple, set)):
+        for item in data:
+            if isinstance(item, _ProcessGraphAbstraction) and item.connection:
+                connections.add(item.connection)
+
+    return connections